change black->transparent in a few templates
These days, the templates are things we're using to draw, not things
we're using to recognize. So where previously we wanted black in places
where nothing got drawn, because that was almost always the thing you
saw there and we wanted to reward places that matched, now we want to
not overwrite places we shouldn't. So this commit changes black to
transparent in a few key templates.

It's possible we may need to change black to transparent in more
templates in the future; this isn't attempting to be a
carefully-thought-out complete commit. It only fixes the one problem
we're seeing right now, which is that when controlled vertical pills are
half out-of-bounds at the top of the bottle, they aren't letting the
bottle graphics through.